Removed the console.log call from the songAdderController.js
Updated the getLocalData, the getAllSongs which was before called with an object as a parameter now use the catName stream as it only parameter. The call to addSongs action created now include the catId and offline paramter.

The rms function was calling the toUppercase function on an object which was trowing an error, I updated the bug.