Added Function that will react to new commit change a spawn new server with the new change. The CommitHandler function will check the received data and pull the remote repository and than close the server and spawn a new server.
The CloseServer will close the running server and the database.
The ForkProcess will fork a new server process.

The start.js file will check in the process is a forked one, and if so it will send a started message to the parent and set an event handle on the message event. This will be use to received close message from the parent and to confirm the closing operation and exit the process.