Updated the AddSongDiv Component. There was a typo made while naming the verse Variable.
Updated the StreamList component. The component was making a call to a non existing property(this.text.listText), i modify the call to a correct one (this.listText).

Updated the HTML component. I added the manifest property in the list of object in the props. When the manifest props is given to the HTML component, it'll add a manifest attribute to the html tag pointing to the value of the manifestFile variable.

The componentDidMount method of the HTML componentn will call the registerWorker method