The appState, Provider, createStore, applyMiddleware, saveUiInfo, checkReachability, timeAction, logAction, myThunk, ManageFastAcces object were removed from the import.
The validator, note, lazyGuider were import in the module.

The lazyGuider will lazy load the guider object when invoqued.
The dbChooser was updated to return a promise which will resolve to a database. The file was updated to wait till the database will be avaible before rendering the App somponent.