Updated the removeCategorie,getCategorie to take as paremeter the categorie id. Also updated all the code in the components.jsx file to use correct parameter for those method. Removed the getCategorieByKey method. Removed uneccessary switch block of code in the database files. Added the deleteCategorieSong method, this will get a categorie id as paremeter and will delete all songs with the given categorie id.
Updated the AddCatDiv,AddSongDiv component to use the isAlphaNumeric method of the Validator to valide categorie name and songName. Also updated the Component to use the correct property value to show the wright message to user.

The updateCat function of the AddCatDiv component was updated. The component will first check if the categorie to be updated is also stored in the browser database and if so it'll update that browser database and then the store.

The updateSong function of the AddSongDiv component was updated. The component will first updated the database and then it the operation was successfull it will update the store. Otherwise it'll show an error to the client.

The CatNames component was updated. The component wipe method will first check if the categorie to be deleted is in the browser database and if so it'll first updated the database and then if the operation was successfull update the store.