Updated the OnlineSongs and OfflineSongs component. 	- The OnlineSongs was updated to only show the online tabs when there is a nom empty current categorie name, and the categorie is online.On initialMount the component will show its song content. 	- The OfflineSongs was updated to only show the offline tabs when there was a non empty current categorie name. The component show it content on mount and when the user change a category.
Updated the Controls component to show image icon instead of the M and X character.

Updated the Item component to show the Control component when the item object is not null and the showControl is a function. Also the Download component will be shown if the src and action2 property are defined.

The Downloads Component was updated to check if the download image should be shown on mount and on update for Categorie and Songs related component.

The SongList was updated:
	- The showControl method was added: this will be use to pass information to the Item Component for showing the Control component.
	- The download method was updated to retrieve the catId from the currentCat state object.
	- The insertSong method was updated to add the given categorie to the browser database if it's not already there before doing any song insert.

The Settings Component was updated to remove the Control Component from it children.
The Control Component was completly removed from the File.